lambda/extract_text/requirements.txt
pdfminer.six
pymysql
